requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
fastapi==0.115.0
Flask==3.0.3
frozenlist==1.4.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gritql==0.1.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.0
langchain-cli==0.0.31
langchain-community==0.3.0
langchain-core==0.3.5
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langserve==0.3.0
langsmith==0.1.125
line-bot-sdk==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.47.0
orjson==3.10.7
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.8.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.38.5
tenacity==8.5.0
tiktoken==0.7.0
tomlkit==0.12.5
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.23.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
